Move process() method from api.py to process.py
